Muitas modificacoes

+++ src/main/java/com/bernardpaula/lojavirtual/service/AdministradorService.java
@@ -4,13 +4,17 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort.Direction;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.bernardpaula.lojavirtual.domain.Administrador;
 import com.bernardpaula.lojavirtual.repository.AdministradorRepository;
+import com.bernardpaula.lojavirtual.service.exceptions.DataIntegrityException;
 import com.bernardpaula.lojavirtual.service.exceptions.ObjectNotFoundException;
 
 @Service
@@ -36,7 +40,17 @@ public class AdministradorService {
 	}
 	
 	public void delete(Integer id) {
-		repo.deleteById(id);
+		//repo.deleteById(id);
+		try {
+			Administrador objCurrent = find(id);
+			if(objCurrent != null) {
+				repo.deleteById(id);
+			}else {
+				 throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "Entidade não encontrada!");
+			}
+		}catch(DataIntegrityViolationException e) {
+			throw new DataIntegrityException("Não foi possível excluir Administrador");
+		}
 	}
 	
 	public List<Administrador> findAll() {
